+++ demo/api/src/db/fixtures/generators/file-uploads-fixture.service.ts
@@ -7,13 +7,11 @@ export class FileUploadsFixtureService {
     constructor(private readonly fileUploadsService: FileUploadsService) {}
 
     async generateFileUploads(): Promise<FileUpload[]> {
-        console.log("Generating file uploads...");
-
         const images = ["01.jpg", "02.jpg", "03.jpg", "04.jpg", "05.jpg"];
         const fileUploads: FileUpload[] = [];
 
         for (const image of images) {
-            const file = await createFileUploadInputFromUrl(path.resolve(`./src/db/fixtures/generators/images/${image}`));
+            const file = await createFileUploadInputFromUrl(path.resolve(`./src/db/fixtures/assets/images/${image}`));
             fileUploads.push(await this.fileUploadsService.upload(file));
         }