+++ packages/admin/admin/src/form/fields/CheckboxField.tsx
@@ -1,10 +1,15 @@
+<<<<<<< HEAD
 import { FormControlLabel, type FormControlLabelProps } from "@mui/material";
+=======
+import { FormControlLabel, FormControlLabelProps } from "@mui/material";
+import { ReactNode } from "react";
+>>>>>>> main
 
 import { FinalFormCheckbox, type FinalFormCheckboxProps } from "../Checkbox";
 import { Field, type FieldProps } from "../Field";
 
 export interface CheckboxFieldProps extends FieldProps<string, HTMLInputElement> {
-    fieldLabel?: string;
+    fieldLabel?: ReactNode;
     componentsProps?: {
         formControlLabel?: FormControlLabelProps;
         finalFormCheckbox?: FinalFormCheckboxProps;