+++ packages/admin/admin/src/theme/componentsTheme/MuiDataGrid.tsx
@@ -1,13 +1,16 @@
 import { ArrowDown, ArrowUp, Check, Clear, Close, Delete, MoreVertical, Search } from "@comet/admin-icons";
 import {
     buttonBaseClasses,
+<<<<<<< HEAD:packages/admin/admin/src/theme/componentsTheme/MuiDataGrid.tsx
     buttonClasses,
     formControlClasses,
+=======
+    getSwitchUtilityClass,
+>>>>>>> main:packages/admin/admin-theme/src/componentsTheme/MuiDataGrid.tsx
     iconButtonClasses,
-    inputAdornmentClasses,
     inputBaseClasses,
-    inputClasses,
     inputLabelClasses,
+    nativeSelectClasses,
     svgIconClasses,
     type SvgIconProps,
     TextField,
@@ -31,16 +34,25 @@ const getDensityHeightValue = (density: string | unknown, spacing: Spacing) => {
     }
 };
 
-export const getMuiDataGrid: GetMuiComponentTheme<"MuiDataGrid"> = (component, { palette, shadows, spacing }) => ({
+const filtersLeftSectionWidth = 120;
+const filterDeleteIconSize = 32;
+const filterLeftSectionGap = 5;
+const filterOperatorInputWidth = filtersLeftSectionWidth - filterDeleteIconSize - filterLeftSectionGap;
+
+export const getMuiDataGrid: GetMuiComponentTheme<"MuiDataGrid"> = (component, { palette, shadows, spacing, breakpoints }) => ({
     ...component,
     defaultProps: {
         ...component?.defaultProps,
+<<<<<<< HEAD:packages/admin/admin/src/theme/componentsTheme/MuiDataGrid.tsx
         disableRowSelectionOnClick: true,
         slots: {
             /* @TODO: add FilterPanelAddIcon to display Comet Add Icon once MUI Datagrid is updated to v6 or higher  */
+=======
+        components: {
+>>>>>>> main:packages/admin/admin-theme/src/componentsTheme/MuiDataGrid.tsx
             QuickFilterIcon: Search,
             QuickFilterClearIcon: Clear,
-            FilterPanelDeleteIcon: Delete,
+            FilterPanelDeleteIcon: (props: SvgIconProps) => <Delete {...props} fontSize="medium" />,
             BooleanCellTrueIcon: Check,
             BooleanCellFalseIcon: Close,
             ColumnSortedAscendingIcon: ArrowUp,
@@ -49,8 +61,16 @@ export const getMuiDataGrid: GetMuiComponentTheme<"MuiDataGrid"> = (component, {
             ColumnMenuIcon: (props: SvgIconProps) => <MoreVertical {...props} fontSize="medium" />,
             ...component?.defaultProps?.slots,
         },
+        componentsProps: {
+            ...component?.defaultProps?.componentsProps,
+            baseButton: {
+                color: "info",
+                ...component?.defaultProps?.componentsProps?.baseButton,
+            },
+        },
         localeText: {
             noRowsLabel: GRID_DEFAULT_LOCALE_TEXT.noResultsOverlayLabel,
+            columnsPanelTextFieldLabel: "",
             ...component?.defaultProps?.localeText,
         },
     },
@@ -69,9 +89,41 @@ export const getMuiDataGrid: GetMuiComponentTheme<"MuiDataGrid"> = (component, {
                 },
             },
         },
+<<<<<<< HEAD:packages/admin/admin/src/theme/componentsTheme/MuiDataGrid.tsx
         columnHeader: ({ ownerState }) => ({
             /* !important is required to override inline styles */
             height: `${getDensityHeightValue(ownerState?.density, spacing)} !important`,
+=======
+        panelHeader: {
+            padding: `4px 4px ${spacing(1)} 4px`,
+            borderBottom: `1px solid ${palette.divider}`,
+        },
+        columnsPanel: {
+            padding: 0,
+        },
+        columnsPanelRow: {
+            marginBottom: spacing(2),
+
+            "&:last-child": {
+                marginBottom: 0,
+            },
+
+            [`& .${switchClasses.root}`]: {
+                marginRight: 0,
+            },
+            [`& .${switchClasses.root} .${switchClasses.thumb}`]: {
+                width: 10,
+                height: 10,
+            },
+            [`& .${switchClasses.root} .${switchClasses.switchBase}`]: {
+                padding: 3,
+            },
+            [`& .${switchClasses.root} .${switchClasses.switchBase}.${getSwitchUtilityClass("checked")}`]: {
+                transform: "translateX(20px)",
+            },
+        },
+        columnHeader: {
+>>>>>>> main:packages/admin/admin-theme/src/componentsTheme/MuiDataGrid.tsx
             "&:focus": {
                 outline: "none",
             },
@@ -124,13 +176,8 @@ export const getMuiDataGrid: GetMuiComponentTheme<"MuiDataGrid"> = (component, {
             height: "20px",
             marginRight: "10px",
         },
-        panel: {
-            ["@media (max-width: 900px)"]: {
-                width: "100%",
-                transform: "translate3d(0,0,0)",
-            },
-        },
         panelContent: {
+<<<<<<< HEAD:packages/admin/admin/src/theme/componentsTheme/MuiDataGrid.tsx
             padding: spacing(1, 0),
             [`& .${gridClasses.filterForm}:first-child .${gridClasses.filterFormLogicOperatorInput}`]: {
                 display: "flex",
@@ -139,42 +186,53 @@ export const getMuiDataGrid: GetMuiComponentTheme<"MuiDataGrid"> = (component, {
                 maxHeight: "none",
                 padding: 0,
             },
+=======
+            padding: spacing(4),
+        },
+        paper: {
+            border: `1px solid ${palette.grey[100]}`,
+            boxShadow: shadows[4],
+            borderRadius: 4,
+            maxHeight: "none",
+            flexDirection: "column",
+>>>>>>> main:packages/admin/admin-theme/src/componentsTheme/MuiDataGrid.tsx
         },
         filterForm: {
-            margin: spacing(5, 4, 0, 4),
-            padding: spacing(2, 1),
-            gap: "5px",
-            borderBottom: `1px solid ${palette.grey[50]}`,
-            ["@media (max-width: 900px)"]: {
-                flexDirection: "row",
-                flexWrap: "wrap",
-                margin: spacing(4, 4, 0, 4),
-                gap: 0,
-                padding: 0,
-                paddingBottom: spacing(5),
-                "&:last-child": {
-                    marginBottom: 0,
-                    paddingBottom: 0,
+            flexDirection: "row",
+            flexWrap: "wrap",
+            padding: 0,
+
+            [`${breakpoints.up("md")}`]: {
+                flexWrap: "nowrap",
+                gap: spacing(1),
+            },
+
+            ["&:not(:last-child)"]: {
+                paddingBottom: spacing(4),
+                marginBottom: spacing(4),
+                borderBottom: `1px solid ${palette.divider}`,
+
+                [`${breakpoints.up("md")}`]: {
+                    paddingBottom: spacing(2),
+                    marginBottom: spacing(2),
+                    borderBottomColor: palette.grey[50],
                 },
             },
-            "&:last-child": {
-                border: "none",
+
+            [`&:first-child .${gridClasses.filterFormLinkOperatorInput}`]: {
+                // The first "Operator"-select is fully hidden by default when there is only one filter.
+                // Setting `display: block` makes sure it takes up it's space as if it were visible to prevent the alignment from breaking.
+                // Even though `display: block` is set now, it's still not visible, due to it's default styling of `visibility: hidden`.
+                display: "block",
             },
-            [`.${formControlClasses.root}`]: {
-                marginRight: 0,
-            },
-            [`.${iconButtonClasses.root}`]: {
-                height: 32,
-                width: 32,
-            },
+
             [`.${inputLabelClasses.root}`]: {
-                transform: "translateY(-22px)",
+                position: "static",
+                transform: "none",
                 fontSize: 14,
-                ["@media (max-width: 900px)"]: {
-                    position: "relative",
-                    transform: "unset",
-                },
+                fontWeight: 600,
             },
+<<<<<<< HEAD:packages/admin/admin/src/theme/componentsTheme/MuiDataGrid.tsx
             [`.${inputClasses.root}`]: {
                 marginTop: 0,
             },
@@ -186,64 +244,81 @@ export const getMuiDataGrid: GetMuiComponentTheme<"MuiDataGrid"> = (component, {
             ["@media (max-width: 900px)"]: {
                 padding: spacing(2, 1),
                 width: "27.2%",
+=======
+
+            [`.${nativeSelectClasses.select}`]: {
+                whiteSpace: "nowrap",
+                textOverflow: "ellipsis",
+>>>>>>> main:packages/admin/admin-theme/src/componentsTheme/MuiDataGrid.tsx
             },
         },
         filterFormDeleteIcon: {
+            width: filterDeleteIconSize,
+            height: filterDeleteIconSize,
+            marginRight: filterLeftSectionGap,
+            marginTop: "auto",
+            marginBottom: 3,
             justifyContent: "center",
 
-            [`& .${svgIconClasses.root}`]: {
-                width: 16,
-                height: 16,
+            [`${breakpoints.up("md")}`]: {
+                marginRight: 0,
             },
 
-            ["@media (max-width: 900px)"]: {
-                padding: spacing(2, 1),
-                alignItems: "flex-start",
-                justifyContent: "flex-end",
-                width: "11.1%",
+            [`& > .${iconButtonClasses.root}`]: {
+                height: "100%",
             },
         },
-        panelFooter: {
-            borderTop: `1px solid ${palette.grey[100]}`,
-            padding: "7px 0",
-            [`.${buttonClasses.root}`]: {
-                color: palette.primary.main,
-            },
-            ["@media (max-width: 900px)"]: {
-                justifyContent: "center",
-                boxShadow: shadows[4],
+        filterFormLinkOperatorInput: {
+            width: filterOperatorInputWidth,
+            marginRight: 0,
+
+            [`${breakpoints.up("md")}`]: {
+                width: 80,
             },
         },
         filterFormColumnInput: {
-            marginRight: spacing(4),
+            width: `calc(100% - ${filtersLeftSectionWidth}px)`,
+            paddingLeft: spacing(2),
+            boxSizing: "border-box",
 
-            ["@media (max-width: 900px)"]: {
-                padding: spacing(2, 1),
-                width: "61.6%",
+            [`${breakpoints.up("md")}`]: {
+                width: 199,
+                paddingLeft: 0,
             },
         },
         filterFormOperatorInput: {
-            margin: spacing(0, 4, 0, 0),
+            marginTop: spacing(3),
+            flexBasis: filterOperatorInputWidth,
+            flexGrow: 1,
 
-            ["@media (max-width: 900px)"]: {
-                padding: spacing(2, 1),
-                width: "38.3%",
+            [`${breakpoints.up("md")}`]: {
+                marginTop: 0,
+                width: 110,
             },
         },
         filterFormValueInput: {
-            ["@media (max-width: 900px)"]: {
-                padding: spacing(2, 1),
-                width: "61.6%",
+            width: `calc(100% - ${filtersLeftSectionWidth}px)`,
+            paddingLeft: spacing(2),
+            boxSizing: "border-box",
+            marginTop: spacing(3),
+
+            [`${breakpoints.up("md")}`]: {
+                width: 199,
+                paddingLeft: 0,
+                marginTop: 0,
+            },
+
+            "&:empty": {
+                display: "none", // Make space for `filterFormOperatorInput` to expand and take up the full width
+            },
+
+            [`& .${inputBaseClasses.root}`]: {
+                marginTop: 0,
             },
         },
-        paper: {
-            boxShadow: shadows[4],
-            border: `1px solid ${palette.divider}`,
-            borderRadius: "4px",
-            ["@media (max-width: 900px)"]: {
-                height: "100%",
-                maxHeight: "none",
-            },
+        panelFooter: {
+            padding: spacing(2),
+            borderTop: `1px solid ${palette.divider}`,
         },
         // @ts-expect-error This key exists but is missing in the types.
         toolbarQuickFilter: {